types of RANGECHECK arguments adjusted.

+++ src/structures/src/String/strdrop.c
@@ -13,7 +13,7 @@ string strdrop(string old, int n)
 {
   string new;
   
-  RANGECHECK(n, 0, strlen(old), old);
+  RANGECHECK((size_t)n, 0, strlen(old), old);
   
   new=(string)SAC_MALLOC(strlen(old)-n+1);
   strcpy(new, old+n);
